Add files via upload
Adding 10 new events (9 cc numu, 1 ccpi0). Each event has both a tracks and a spacepoints version that contains only the pandoraCosmic tracks or spacepoints products. The pytfilt.py script is the filter script and it is run via the runs.sh shell script (which contains the links to argo json outputs that are retrieved via curl).